+++ src/WSEdit/Data.hs
@@ -108,7 +108,7 @@ import qualified WSEdit.Buffer as B
 
 -- | Version number constant.
 version :: String
-version = "1.2.3.4"
+version = "1.2.3.5"
 
 -- | Upstream URL.
 upstream :: String